Shell: Move escape_token* functions to their own header
Affected functions are:
- escape_token_for_double_quotes()
- escape_token_for_single_quotes()
- escape_token()
- unescape_token()
