+++ Userland/Shell/Shell.h
@@ -6,6 +6,7 @@
 
 #pragma once
 
+#include "Escape.h"
 #include "Job.h"
 #include "Parser.h"
 #include <AK/CircularQueue.h>
@@ -152,18 +153,6 @@ public:
 
     [[nodiscard]] Frame push_frame(String name);
     void pop_frame();
-
-    static String escape_token_for_double_quotes(StringView token);
-    static String escape_token_for_single_quotes(StringView token);
-    static String escape_token(StringView token);
-    static String unescape_token(StringView token);
-    enum class SpecialCharacterEscapeMode {
-        Untouched,
-        Escaped,
-        QuotedAsEscape,
-        QuotedAsHex,
-    };
-    static SpecialCharacterEscapeMode special_character_escape_mode(u32 c);
 
     static bool is_glob(StringView);
     static Vector<StringView> split_path(StringView);
@@ -378,16 +367,16 @@ inline size_t find_offset_into_node(const String& unescaped_text, size_t escaped
             if (offset == escaped_offset)
                 return unescaped_offset;
 
-            switch (Shell::special_character_escape_mode(c)) {
-            case Shell::SpecialCharacterEscapeMode::Untouched:
+            switch (special_character_escape_mode(c)) {
+            case SpecialCharacterEscapeMode::Untouched:
                 break;
-            case Shell::SpecialCharacterEscapeMode::Escaped:
+            case SpecialCharacterEscapeMode::Escaped:
                 ++offset; // X -> \X
                 break;
-            case Shell::SpecialCharacterEscapeMode::QuotedAsEscape:
+            case SpecialCharacterEscapeMode::QuotedAsEscape:
                 offset += 3; // X -> "\Y"
                 break;
-            case Shell::SpecialCharacterEscapeMode::QuotedAsHex:
+            case SpecialCharacterEscapeMode::QuotedAsHex:
                 if (c > NumericLimits<u8>::max())
                     offset += 11; // X -> "\uhhhhhhhh"
                 else
